# ORC-1150: [C++] Optimize RowReaderImpl::computeBatchSize() by pre-computation