--- a/c++/test/TestReader.cc
+++ b/c++/test/TestReader.cc
@@ -60,53 +60,55 @@ namespace orc {
     uint64_t rowsInCurrentStripe = 100 * 8 + 50;
     std::vector<bool> includedRowGroups =
       { false, false, true, true, false, false, true, true, false };
+    std::vector<uint64_t> nextSkippedRows =
+      { 0, 0, 400, 400, 0, 0, 800, 800, 0 };
 
     EXPECT_EQ(0, RowReaderImpl::computeBatchSize(
-      1024, 0, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      1024, 0, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(0, RowReaderImpl::computeBatchSize(
-      1024, 50, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      1024, 50, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(200, RowReaderImpl::computeBatchSize(
-      1024, 200, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      1024, 200, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(150, RowReaderImpl::computeBatchSize(
-      1024, 250, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      1024, 250, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(0, RowReaderImpl::computeBatchSize(
-      1024, 550, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      1024, 550, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(100, RowReaderImpl::computeBatchSize(
-      1024, 700, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      1024, 700, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(50, RowReaderImpl::computeBatchSize(
-      50, 700, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      50, 700, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(0, RowReaderImpl::computeBatchSize(
-      50, 810, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      50, 810, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(0, RowReaderImpl::computeBatchSize(
-      50, 900, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      50, 900, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
   }
 
   TEST(TestRowReader, advanceToNextRowGroup) {
     uint64_t rowIndexStride = 100;
     uint64_t rowsInCurrentStripe = 100 * 8 + 50;
-    std::vector<bool> includedRowGroups =
-      { false, false, true, true, false, false, true, true, false };
+    std::vector<uint64_t> nextSkippedRows =
+      { 0, 0, 400, 400, 0, 0, 800, 800, 0 };
 
     EXPECT_EQ(200, RowReaderImpl::advanceToNextRowGroup(
-      0, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      0, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(200, RowReaderImpl::advanceToNextRowGroup(
-      150, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      150, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(250, RowReaderImpl::advanceToNextRowGroup(
-      250, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      250, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(350, RowReaderImpl::advanceToNextRowGroup(
-      350, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      350, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(350, RowReaderImpl::advanceToNextRowGroup(
-      350, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      350, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(600, RowReaderImpl::advanceToNextRowGroup(
-      500, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      500, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(699, RowReaderImpl::advanceToNextRowGroup(
-      699, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      699, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(799, RowReaderImpl::advanceToNextRowGroup(
-      799, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      799, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(850, RowReaderImpl::advanceToNextRowGroup(
-      800, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      800, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
     EXPECT_EQ(850, RowReaderImpl::advanceToNextRowGroup(
-      900, rowsInCurrentStripe, rowIndexStride, includedRowGroups));
+      900, rowsInCurrentStripe, rowIndexStride, nextSkippedRows));
   }
 
   void CheckFileWithSargs(const char* fileName, const char* softwareVersion) {